+++ src/core/data/pic_electrons/electron_population.hpp
@@ -1,76 +1,64 @@
-#ifndef PHARE_ION_POPULATION_HPP
-#define PHARE_ION_POPULATION_HPP
+#ifndef PHARE_ELECTRON_POPULATION_HPP
+#define PHARE_ELECTRON_POPULATION_HPP
 
 #include <memory>
 #include <sstream>
 #include <stdexcept>
 #include <string>
-#include <tuple>
-#include <array>
-
 
 #include "core/def.hpp"
 #include "core/hybrid/hybrid_quantities.hpp"
 #include "initializer/data_provider.hpp"
-#include "particle_pack.hpp"
-#include "core/data/pic_electrons/particle_population.hpp"
+#include "core/data/ions/ion_population/particle_pack.hpp"
+#include "particle_population.hpp"
 
 namespace PHARE::core
 {
     template<typename ParticleArray, typename VecField, typename TensorField, typename GridLayout>
-    class IonPopulation: public ParticlePopulation<ParticleArray, VecField, TensorField, 
-                                                   GridLayout, IonPopulation>
+    class ElectronPopulation : public ParticlePopulation<ParticleArray, VecField, TensorField, 
+                                                         GridLayout, ElectronPopulation>
     {
     public:
         using field_type                       = typename VecField::field_type;
         using particle_array_type              = ParticleArray;
         using particle_resource_type           = ParticlesPack<ParticleArray>;
         using vecfield_type                    = VecField;
-        using tensorfield_type                 = TensorField;
 
         using Super = ParticlePopulation<ParticleArray, VecField, TensorField, GridLayout, 
-                                         IonPopulation>;
+                                         ElectronPopulation>;
         using Super::name_;
 
-
-        IonPopulation(initializer::PHAREDict const& initializer)
-            : Super{initializer}
-            , flux_{name_ + "_flux", HybridQuantity::Vector::V}
-            , momentumTensor_{name_ + "_momentumTensor", HybridQuantity::Tensor::M}
+        ElectronPopulation(initializer::PHAREDict const& initializer)
+            , flux_{name_ + "_flux", HybridQuantity::Vector::Ve}
         {
-        }
-
-
-        NO_DISCARD bool isUsable() const override
-        {
-            return particles_ != nullptr && rho_ != nullptr && flux_.isUsable()
-                   && momentumTensor_.isUsable();
-        }
-
-
-        NO_DISCARD bool isSettable() const override
-        {
-            return particles_ == nullptr && rho_ == nullptr && flux_.isSettable()
-                   && momentumTensor_.isSettable();
-        }
-
+        } 
 
 
         NO_DISCARD VecField const& flux() const { return flux_; }
         NO_DISCARD VecField& flux() { return flux_; }
-
-        NO_DISCARD TensorField const& momentumTensor() const { return momentumTensor_; }
-        NO_DISCARD TensorField& momentumTensor() { return momentumTensor_; }
 
         NO_DISCARD field_type const& rho() const override { return *rho_; }
 
         NO_DISCARD ParticlesPack<ParticleArray> const* getParticlesPtr() const override { 
             return particles_; }
 
+
+        NO_DISCARD bool isUsable() const
+        {
+            return particles_ != nullptr && rho_ != nullptr && flux_.isUsable();
+        }
+
+
+        NO_DISCARD bool isSettable() const
+        {
+            return particles_ == nullptr && rho_ == nullptr && flux_.isSettable();
+        }
+
+
+
         //-------------------------------------------------------------------------
         //                  start the ResourcesUser interface
         //-------------------------------------------------------------------------
-
 
 
         using typename Super::MomentProperties;
@@ -78,15 +66,7 @@ namespace PHARE::core
 
         NO_DISCARD MomentProperties getFieldNamesAndQuantities() const
         {
-            return {{{name_ + "_rho", HybridQuantity::Scalar::rho}}};
-        }
-
-
-
-
-        NO_DISCARD auto getCompileTimeResourcesUserList()
-        {
-            return std::forward_as_tuple(flux_, momentumTensor_);
+            return {{{name_ + "_rho", HybridQuantity::Scalar::rhoE}}};
         }
 
 
@@ -102,25 +82,28 @@ namespace PHARE::core
         void setBuffer(std::string const& bufferName, field_type* field)
         {
             if (bufferName == name_ + "_rho")
-            {
                 rho_ = field;
-            }
             else
-            {
                 throw std::runtime_error("Error - invalid density buffer name");
-            }
         }
+
+
+
+        NO_DISCARD auto getCompileTimeResourcesUserList()
+        {
+            return std::forward_as_tuple(flux_);
+        }
+        
 
         //-------------------------------------------------------------------------
         //                  ends the ResourcesUser interface
         //-------------------------------------------------------------------------
 
-        std::string getParticleName() const override { return "Ion"; }
+        std::string getParticleName() const override { return "Electron"; }
 
 
     private:
         VecField flux_;
-        TensorField momentumTensor_;
         field_type* rho_{nullptr};
         ParticlesPack<ParticleArray>* particles_{nullptr};
     };
