+++ tests/core/data/electron_population/CMakeLists.txt
@@ -1,0 +1,20 @@
+cmake_minimum_required (VERSION 3.20.1)
+
+project(test-electron-population)
+
+set(SOURCES test_electron_population.cpp)
+
+add_executable(${PROJECT_NAME} ${SOURCES})
+
+target_include_directories(${PROJECT_NAME} PRIVATE
+  $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/src>
+  ${GTEST_INCLUDE_DIRS}
+  )
+
+target_link_libraries(${PROJECT_NAME} PRIVATE
+  phare_core
+  ${GTEST_LIBS})
+
+add_no_mpi_phare_test(${PROJECT_NAME} ${CMAKE_CURRENT_BINARY_DIR})
+
+
